arv_python/requirements.txt
openpyxl>=3.0.0
python-docx>=0.8.11
matplotlib>=3.5.0
# Для работы с Rastr (ПК RUSTab) под Windows:
# pywin32>=305
